extract in memory Besu blockchain
