+++ src/main/kotlin/org/web3j/evm/EmbeddedWeb3jService.kt
@@ -187,8 +187,8 @@ class EmbeddedWeb3jService(configuration: Configuration, operationTracer: Operat
 
     private fun ethCall(params: List<*>): Response<String> {
         val transaction = params[0] as Transaction
-        val defaultBlockParameter = params[1].toString()
-        val result = embeddedEthereum.ethCall(transaction, defaultBlockParameter)
+        // val defaultBlockParameter = params[1].toString()
+        val result = embeddedEthereum.ethCall(transaction)
 
         return object : EthCall() {
             override fun getResult(): String {
